make backlight configurable
This makes the backlight values configurable. it introduces a
curvefactor that modifies the curve of the logarithmic steps. Human
brightness perception is not linear, but it's not entirely logarithmic
either, at least not on the author's laptop. The author finds that a
curvefactor of 0.5 makes the brightness changes appear completely
linear.